requirements.txt
python-dotenv>=1.0
docxtpl>=0.16
python-docx>=0.8.11
docx2pdf>=0.1.8 ; sys_platform != "linux"
